Refactoring: move subscription's sessions rebalance part of the `ZkSubscriptionClient`
