+++ core-common/src/main/java/org/zalando/nakadi/service/subscription/zk/SubscriptionRebalancer.java
@@ -1,4 +1,4 @@
-package org.zalando.nakadi.service.subscription;
+package org.zalando.nakadi.service.subscription.zk;
 
 import com.google.common.collect.Lists;
 import org.zalando.nakadi.domain.EventTypePartition;
