+++ core-common/src/test/java/org/zalando/nakadi/service/subscription/zk/SubscriptionRebalancerTest.java
@@ -1,4 +1,4 @@
-package org.zalando.nakadi.service.subscription;
+package org.zalando.nakadi.service.subscription.zk;
 
 import com.google.common.collect.ImmutableList;
 import org.junit.Test;
